requirements.txt
python-telegram-bot
gpio